add in completed component
