+++ src/state/tasks/reducers.js
@@ -33,13 +33,13 @@ const tasksSlice = createSlice({
         id: 'ghi',
       },
       {
-        label: 'Add in ids for each item being created',
+        label: 'Reset selected items when you click change category',
         date: '2024-01-05',
         details: '',
         category: 'coding',
         active: false,
         status: 'incomplete',
-        id: 'jkl',
+        id: 'ghi',
       },
     ],
   },
